Sync to upstream/release/698 (#2070)
# General

Another week, another release. Happy Halloween! 🎃

## Require
- Integrate support for `.config.luau` files into `Luau.Require` as part
of the [Luau-syntax configuration files
RFC](https://rfcs.luau.org/config-luauconfig.html).
  - `is_config_present` has been replaced with `get_config_status`.
- `get_luau_config_timeout` has been added to configure extraction
timeouts at runtime.
- Enable support for `.config.luau` files in `luau` and `luau-analyze`.
- Merge the `Luau.RequireNavigator` static library into `Luau.Require`.

## Analysis
- Add fuel-based limits for normalization: normalization will now take a
set number of steps (the "fuel") and stop when we hit a certain number
of steps ("run out of fuel"). This is in lieu of trying to check static
limits on the size of its inputs. This may result in seeing more "code
too complex" errors but should dramatically reduce the number of cases
where Luau becomes unresponsive and uses several gigabytes of memory. We
plan to tune this limit over time to find the right balance between
responsiveness and completeness.
- Fix a case where refining a variable with the type `any` would cause
it to become `unknown` instead due to normalization not preserving `any`
in the top type position.
- Improve the wording of type errors in the new non-strict mode.
- Fix #1910.
- Fix #2065.
- Fix #1483.
- Fix #2018.

## Miscellaneous
- Introduce `Luau::overloaded` to facilitate interacting with
`Luau::Variant`.
- Add type checking support to the [Luau demo](https://luau.org/demo)!

---

Co-authored-by: Annie Tang <annietang@roblox.com>
Co-authored-by: Ariel Weiss <arielweiss@roblox.com>
Co-authored-by: Hunter Goldstein <hgoldstein@roblox.com>
Co-authored-by: Ilya Rezvov <irezvov@roblox.com>
Co-authored-by: Sora Kanosue <skanosue@roblox.com>
Co-authored-by: Varun Saini <vsaini@roblox.com>