Explicit type instantiations (`f<<T>>`) (#1980)
This adds in support for [Explicit type parameter
instantiation](https://rfcs.luau.org/explicit-type-parameter-instantiation.html).

This PR is still a work in progress, but some important notes:

1. Metatables with `__call` are **not** supported. While `t<<A>>()` is
obvious what it should do with `__call`, it's not obvious what `t<<A>>`
on its own would be.
2. Intersection types are not supported at the moment either.

Both of these are possible to bring later.